+++ examples/src/test/java/org/reactivestreams/example/unicast/RangePublisherTest.java
@@ -9,9 +9,10 @@
  * work. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.*
  ************************************************************************/
 
-package org.reactivestreams.example.lazycast;
+package org.reactivestreams.example.unicast;
 
 import org.reactivestreams.Publisher;
+import org.reactivestreams.example.unicast.RangePublisher;
 import org.reactivestreams.tck.*;
 
 public class RangePublisherTest extends PublisherVerification<Integer> {
